requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
starlette>=0.27.0
pydantic>=2.0.0
requests>=2.31.0
python-multipart>=0.0.6
pillow>=10.0.0
openai>=1.0.0
dashscope>=1.14.0
python-dotenv>=1.0.0
pytest>=8.0.0
httpx>=0.27.0
